Move useEventHandler hook to the helpers file
